cmms-backend/requirements.txt
# FastAPI & Server
fastapi>=0.110
uvicorn[standard]>=0.20

# Database & ORM
sqlalchemy>=2.0
psycopg2-binary>=2.9

# Pydantic (Model Validation)
pydantic>=2.7
pydantic-settings>=2.2
pydantic[email]
email-validator

# Authentication & Security
python-jose>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6

# Migrations (Optional but recommended)
alembic>=1.13

# Logging
loguru>=0.7
